move spatial defaults to `constants-internal`
