+++ packages/utils/spatial-utils/src/index.js
@@ -6,6 +6,8 @@ export {
   DEFAULT_RASTER_DOMAIN_TYPE,
   DEFAULT_RASTER_LAYER_PROPS,
   DEFAULT_LAYER_TYPE_ORDERING,
+} from '@vitessce/constants-internal';
+export {
   square,
   initializeLayerChannels,
   initializeRasterLayersAndChannels,
